avg_marks/requirements.txt
django
djangorestframework
django-filters?